ENHANCE: Not to close and reopen pool connections #59
